Redefine TextArea#getCharacterAt() as getInsertionPoint(); remove some redundant casts in HexUtils and MD5.
git-svn-id: https://svn.apache.org/repos/asf/incubator/pivot/trunk@828850 13f79535-47bb-0310-9956-ffa450edef68

+++ web/src/org/apache/pivot/web/MD5.java
@@ -131,8 +131,8 @@ public final class MD5 {
         char[] buffer = new char[MD5_DIGEST_LENTGH_IN_BYTES * 2];
 
         for (int i = 0; i < MD5_DIGEST_LENTGH_IN_BYTES; i++) {
-            int low = (int) (binaryData[i] & 0x0f);
-            int high = (int) ((binaryData[i] & 0xf0) >> 4);
+            int low = binaryData[i] & 0x0f;
+            int high = (binaryData[i] & 0xf0) >> 4;
 
             buffer[i * 2] = HexUtils.hexadecimal[high];
             buffer[i * 2 + 1] = HexUtils.hexadecimal[low];
